+++ .changeset/thirty-beers-occur.md
@@ -1,0 +1,5 @@
+---
+'@igloo-ui/helper-text': minor
+---
+
+Rename component `Message` by `HelperText`
