+++ packages/HelperText/README.md
@@ -1,0 +1,19 @@
+# @igloo-ui/helper-text
+
+Welcome to my `HelperText` component.
+
+## Installation
+
+```sh
+npm i @igloo-ui/helper-text
+# or with yarn
+yarn add @igloo-ui/helper-text
+```
+
+Then to use the component in your code just import it!
+
+```js
+import HelperText from '@igloo-ui/helper-text';
+// and with css-modules
+import '@igloo-ui/HelperText/dist/helper-text.css';
+```
