--- packages/Message/README.md
@@ -1,19 +1,0 @@
-# @igloo-ui/message
-
-Welcome to my `Message` component.
-
-## Installation
-
-```sh
-npm i @igloo-ui/message
-# or with yarn
-yarn add @igloo-ui/message
-```
-
-Then to use the component in your code just import it!
-
-```js
-import Message from '@igloo-ui/message';
-// and with css-modules
-import '@igloo-ui/Message/dist/message.css';
-```
