--- packages/Message/src/message.scss
@@ -1,32 +1,0 @@
-@import '~@igloo-ui/tokens/dist/base10/variables';
-@import '~@igloo-ui/tokens/dist/fonts.css';
-
-:root {
-  /* Default */
-  --ids-message-font-family: #{$primary-font-family};
-  --ids-message-font-size: #{$font-size-3};
-  --ids-message-font-weight: #{$font-weight-regular};
-  --ids-message-line-height: #{$line-height-xxl};
-  --ids-message-text: #{$grey-800};
-
-  /* Error */
-  --ids-message-text-error: #{$coral-500};
-}
-
-.ids-message {
-  display: inline-flex;
-  align-items: center;
-  font-family: var(--ids-message-font-family);
-  font-size: var(--ids-message-font-size);
-  font-weight: var(--ids-message-font-weight);
-  line-height: var(--ids-message-line-height);
-  color: var(--ids-message-text);
-
-  &--error {
-    color: var(--ids-message-text-error);
-  }
-
-  > .ids-message__icon {
-    margin-right: $space-1;
-  }
-}
